requirements.txt
fastapi
mangum
uvicorn
web3
python-dotenv
py-solc-x
pydantic